Moved ServiceConfig class into the service-registration project
ServiceConfig is a class needed only during development, and in fact it is used
only by the service auto-registering logic taking care of automatically
registering services to the development service locator.

This was one class it was forgotten by the refactoring carried out in
f177a370d1cc07bac30f0ab4cdcbad9098c1cbf4.
