+++ server/src/main/resources/reference.conf
@@ -1,4 +1,2 @@
-play.modules.enabled += com.lightbend.lagom.internal.server.ServerModule
-
 # Possibility to disable the built-in status endpoint
 lagom.status-endpoint.enabled = on
